+++ pdf/lib/src/pdf/obj/array.dart
@@ -19,6 +19,7 @@ import '../format/array.dart';
 import 'object.dart';
 
 /// An array object
+@Deprecated('Use PdfObject<PdfArray> instead')
 class PdfArrayObject extends PdfObject<PdfArray> {
   /// Creates an array object
   PdfArrayObject(