+++ pdf/lib/src/pdf/obj/encryption.dart
@@ -17,13 +17,15 @@
 import 'dart:typed_data';
 
 import '../document.dart';
+import '../format/dict.dart';
 import '../format/object_base.dart';
-import 'object_dict.dart';
+import 'object.dart';
 
 /// Encryption object
-abstract class PdfEncryption extends PdfObjectDict {
+abstract class PdfEncryption extends PdfObject<PdfDict> {
   /// Creates an encryption object
-  PdfEncryption(PdfDocument pdfDocument) : super(pdfDocument);
+  PdfEncryption(PdfDocument pdfDocument)
+      : super(pdfDocument, params: PdfDict());
 
   /// Encrypt some data
   Uint8List encrypt(Uint8List input, PdfObjectBase object);