+++ pdf/lib/src/pdf/obj/info.dart
@@ -15,43 +15,36 @@
  */
 
 import '../document.dart';
+import '../format/dict.dart';
 import '../format/string.dart';
-import 'object_dict.dart';
+import 'object.dart';
 
 /// Information object
-class PdfInfo extends PdfObjectDict {
+class PdfInfo extends PdfObject<PdfDict> {
   /// Create an information object
-  PdfInfo(PdfDocument pdfDocument,
-      {this.title,
-      this.author,
-      this.creator,
-      this.subject,
-      this.keywords,
-      this.producer})
-      : super(pdfDocument) {
-    if (author != null) {
-      params['/Author'] = PdfString.fromString(author!);
-    }
-    if (creator != null) {
-      params['/Creator'] = PdfString.fromString(creator!);
-    }
-    if (title != null) {
-      params['/Title'] = PdfString.fromString(title!);
-    }
-    if (subject != null) {
-      params['/Subject'] = PdfString.fromString(subject!);
-    }
-    if (keywords != null) {
-      params['/Keywords'] = PdfString.fromString(keywords!);
-    }
-    if (producer != null) {
-      params['/Producer'] = PdfString.fromString('$producer ($_libraryName)');
-    } else {
-      params['/Producer'] = PdfString.fromString(_libraryName);
-    }
-
-    params['/CreationDate'] = PdfString.fromDate(DateTime.now());
-  }
+  PdfInfo(
+    PdfDocument pdfDocument, {
+    this.title,
+    this.author,
+    this.creator,
+    this.subject,
+    this.keywords,
+    this.producer,
+  }) : super(
+          pdfDocument,
+          params: PdfDict({
+            if (author != null) '/Author': PdfString.fromString(author),
+            if (creator != null) '/Creator': PdfString.fromString(creator),
+            if (title != null) '/Title': PdfString.fromString(title),
+            if (subject != null) '/Subject': PdfString.fromString(subject),
+            if (keywords != null) '/Keywords': PdfString.fromString(keywords),
+            if (producer != null)
+              '/Producer': PdfString.fromString('$producer ($_libraryName)')
+            else
+              '/Producer': PdfString.fromString(_libraryName),
+            '/CreationDate': PdfString.fromDate(DateTime.now()),
+          }),
+        );
 
   static const String _libraryName = 'https://github.com/DavBfr/dart_pdf';