+++ pdf/lib/src/pdf/obj/shading.dart
@@ -17,16 +17,17 @@
 import '../document.dart';
 import '../format/array.dart';
 import '../format/bool.dart';
+import '../format/dict.dart';
 import '../format/name.dart';
 import '../format/num.dart';
 import '../point.dart';
 import '../rect.dart';
 import 'function.dart';
-import 'object_dict.dart';
+import 'object.dart';
 
 enum PdfShadingType { axial, radial }
 
-class PdfShading extends PdfObjectDict {
+class PdfShading extends PdfObject<PdfDict> {
   PdfShading(
     PdfDocument pdfDocument, {
     required this.shadingType,
@@ -38,7 +39,7 @@ class PdfShading extends PdfObjectDict {
     this.boundingBox,
     this.extendStart = false,
     this.extendEnd = false,
-  }) : super(pdfDocument);
+  }) : super(pdfDocument, params: PdfDict());
 
   /// Name of the Shading object
   String get name => '/S$objser';