+++ pdf/lib/src/pdf/obj/signature.dart
@@ -18,9 +18,9 @@ import 'dart:typed_data';
 
 import '../document.dart';
 import '../format/dict.dart';
+import '../format/name.dart';
 import '../format/stream.dart';
 import 'object.dart';
-import 'object_dict.dart';
 import 'object_stream.dart';
 
 /// Signature flags
@@ -35,7 +35,7 @@ enum PdfSigFlags {
   appendOnly,
 }
 
-class PdfSignature extends PdfObjectDict {
+class PdfSignature extends PdfObject<PdfDict> {
   PdfSignature(
     PdfDocument pdfDocument, {
     required this.value,
@@ -43,7 +43,12 @@ class PdfSignature extends PdfObjectDict {
     List<Uint8List>? crl,
     List<Uint8List>? cert,
     List<Uint8List>? ocsp,
-  }) : super(pdfDocument, type: '/Sig') {
+  }) : super(
+          pdfDocument,
+          params: PdfDict({
+            '/Type': const PdfName('/Sig'),
+          }),
+        ) {
     if (crl != null) {
       for (final o in crl) {
         this.crl.add(PdfObjectStream(pdfDocument)..buf.putBytes(o));